Tweaks to get rid of compiler warnings on 64-bit CPU's.  Fixing up mostly casts to/from pointers and int's using some macros from the FreeCIV project to help us.

+++ src/maps.c
@@ -1,5 +1,5 @@
 /* -*- c-basic-offset: 4; indent-tabs-mode: nil -*-
- * $Id: maps.c,v 1.415 2004/09/10 03:37:33 we7u Exp $
+ * $Id: maps.c,v 1.416 2004/10/02 18:08:20 we7u Exp $
  *
  * XASTIR, Amateur Station Tracking and Information Reporting
  * Copyright (C) 1999,2000  Frank Giannandrea
@@ -244,7 +244,10 @@ int utm_grid_clear(int do_alloc)
                 utm_grid.zone[i].col[j].point = calloc(UTM_GRID_DEF_NALLOCED, sizeof(XPoint));
                 if (!utm_grid.zone[i].col[j].point) {
                     fprintf(stderr,"calloc(%d, %d) for z=%d col=%d FAILED!\n",
-                           UTM_GRID_DEF_NALLOCED, sizeof(XPoint), i, j);
+                        UTM_GRID_DEF_NALLOCED,
+                        (int)sizeof(XPoint),
+                        i,
+                        j);
 //                    abort();  // Causes a segfault
                     return(1);
                 }
@@ -260,7 +263,10 @@ int utm_grid_clear(int do_alloc)
                 utm_grid.zone[i].row[j].point = calloc(UTM_GRID_DEF_NALLOCED, sizeof(XPoint));
                 if (!utm_grid.zone[i].row[j].point) {
                     fprintf(stderr,"calloc(%d, %d) for z=%d row=%d FAILED!\n",
-                           UTM_GRID_DEF_NALLOCED, sizeof(XPoint), i, j);
+                        UTM_GRID_DEF_NALLOCED,
+                        (int)sizeof(XPoint),
+                        i,
+                        j);
 //                    abort();  // Causes a segfault
                     return(1);
                 }
